# packages/settlrl-app/frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Settlrl",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/lobbies": {
   "get": {
    "summary": "List Lobbies",
    "description": "Listed, joinable online lobbies (newest first).",
    "operationId": "list_lobbies_api_lobbies_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/_LobbyListModel"
         },
         "title": "Response List Lobbies Api Lobbies Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Create Lobby",
    "description": "Open a lobby. The host takes seat 0 (online) or every human seat\n(hotseat). Listing or Quick-Match visibility needs a signed-in account.",
    "operationId": "create_lobby_api_lobbies_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "X-Client-Id",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Client-Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_CreateLobbyRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/_CreatedLobbyModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}": {
   "get": {
    "summary": "Get Lobby",
    "operationId": "get_lobby_api_lobbies__lobby_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/_LobbyModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}/events": {
   "get": {
    "summary": "Lobby Events",
    "description": "The requester's lobby snapshot now, then again on every change (joins,\nseat edits, config, chat) and once more carrying ``started_game_id``.",
    "operationId": "lobby_events_api_lobbies__lobby_id__events_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}/configure": {
   "post": {
    "summary": "Configure Lobby",
    "description": "Host-only (seat-0 owner) pre-start config. Changing the player count\nrebuilds the seat list, keeping seat 0 (the host).",
    "operationId": "configure_lobby_api_lobbies__lobby_id__configure_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ConfigureLobbyRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/_LobbyModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}/seats": {
   "post": {
    "summary": "Set Seat",
    "description": "Host-only: retarget a seat \u2014 a bot kind, or open it (online) / take it\n(hotseat) as human. The host can change its own seats (the whole table in\na hotseat) but not seat 0 (its anchor) nor a seat another player holds.",
    "operationId": "set_seat_api_lobbies__lobby_id__seats_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_SeatRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/_LobbyModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}/join": {
   "post": {
    "summary": "Join Lobby",
    "description": "Take an open human seat in an online lobby.",
    "operationId": "join_lobby_api_lobbies__lobby_id__join_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Client-Id",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Client-Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_JoinRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/_CreatedLobbyModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}/leave": {
   "post": {
    "summary": "Leave Lobby",
    "description": "The host (seat 0) closes the whole lobby; anyone else frees their\nseat. 403 if you hold no seat here.",
    "operationId": "leave_lobby_api_lobbies__lobby_id__leave_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}/chat": {
   "post": {
    "summary": "Lobby Chat",
    "operationId": "lobby_chat_api_lobbies__lobby_id__chat_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ChatRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/_LobbyModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/lobbies/{lobby_id}/start": {
   "post": {
    "summary": "Start Lobby",
    "description": "Host-only: materialise the lobby into a game. 409 while any human seat\nis still open (start a table only once every seat is decided). A queued\nbody with a place in line when the server is at its game cap.",
    "operationId": "start_lobby_api_lobbies__lobby_id__start_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "lobby_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Lobby Id"
      }
     },
     {
      "name": "X-Seat-Tokens",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Seat-Tokens"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_StartRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/_StartedModel"
          },
          {
           "$ref": "#/components/schemas/_QueuedModel"
          }
         ],
         "title": "Response Start Lobby Api Lobbies  Lobby Id  Start Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "_ChatRequest": {
    "properties": {
     "text": {
      "type": "string",
      "title": "Text"
     },
     "player": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Player"
     }
    },
    "type": "object",
    "required": [
     "text"
    ],
    "title": "_ChatRequest"
   },
...
   "_CreateLobbyRequest": {
    "properties": {
     "mode": {
      "type": "string",
      "enum": [
       "hotseat",
       "online"
      ],
      "title": "Mode",
      "default": "online"
     },
     "n_players": {
      "type": "integer",
      "enum": [
       2,
       4
      ],
      "title": "N Players",
      "default": 4
     },
     "seed": {
      "type": "integer",
      "title": "Seed",
      "default": 0
     },
     "number_placement": {
      "type": "string",
      "enum": [
       "random",
       "spiral"
      ],
      "title": "Number Placement",
      "default": "random"
     },
     "victory_points_to_win": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Victory Points To Win"
     },
     "listed": {
      "type": "boolean",
      "title": "Listed",
      "default": false
     },
     "searchable": {
      "type": "boolean",
      "title": "Searchable",
      "default": false
     }
    },
    "type": "object",
    "title": "_CreateLobbyRequest"
   },
   "_CreatedLobbyModel": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "tokens": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Tokens"
     }
    },
    "type": "object",
    "required": [
     "id",
     "tokens"
    ],
    "title": "_CreatedLobbyModel"
   },
...
   "_JoinRequest": {
    "properties": {
     "seat": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seat"
     }
    },
    "type": "object",
    "title": "_JoinRequest"
   },
...
   "_LobbyListModel": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "n_players": {
      "type": "integer",
      "title": "N Players"
     },
     "number_placement": {
      "type": "string",
      "title": "Number Placement"
     },
     "open_seats": {
      "type": "integer",
      "title": "Open Seats"
     },
     "searchable": {
      "type": "boolean",
      "title": "Searchable"
     },
     "created_at": {
      "type": "number",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "n_players",
     "number_placement",
     "open_seats",
     "searchable",
     "created_at"
    ],
    "title": "_LobbyListModel"
   },
   "_LobbyModel": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "mode": {
      "type": "string",
      "title": "Mode"
     },
     "seed": {
      "type": "integer",
      "title": "Seed"
     },
     "number_placement": {
      "type": "string",
      "title": "Number Placement"
     },
     "n_players": {
      "type": "integer",
      "title": "N Players"
     },
     "victory_points_to_win": {
      "type": "integer",
      "title": "Victory Points To Win"
     },
     "kinds": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Kinds"
     },
     "seats_claimed": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Seats Claimed"
     },
     "seat_names": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ]
      },
      "type": "array",
      "title": "Seat Names"
     },
     "your_seats": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Your Seats"
     },
     "listed": {
      "type": "boolean",
      "title": "Listed"
     },
     "searchable": {
      "type": "boolean",
      "title": "Searchable"
     },
     "ready": {
      "type": "boolean",
      "title": "Ready"
     },
     "started_game_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started Game Id"
     },
     "board": {
      "$ref": "#/components/schemas/BoardModel"
     },
     "chat": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Chat"
     }
    },
    "type": "object",
    "required": [
     "id",
     "mode",
     "seed",
     "number_placement",
     "n_players",
     "victory_points_to_win",
     "kinds",
     "seats_claimed",
     "seat_names",
     "your_seats",
     "listed",
     "searchable",
     "ready",
     "started_game_id",
     "board",
     "chat"
    ],
    "title": "_LobbyModel",
    "description": "One lobby as a requester sees it: its config, seats, and preview board."
   },
...
   "_SeatRequest": {
    "properties": {
     "seat": {
      "type": "integer",
      "title": "Seat"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     }
    },
    "type": "object",
    "required": [
     "seat",
     "kind"
    ],
    "title": "_SeatRequest"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "api/auth/login"
     }
    }
   }
  }
 }
}